refactor: Reorganize codebase structure for improved clarity
This PR improves codebase organization through focused renaming and restructuring:

## Changes Made

### 1. Rename Files for Clarity
- `processing/validators.py` → `processing/validation_utils.py` (internal validation utilities)
- `dsl/simple_type.py` → `dsl/primitives.py` (clearer name for primitive type handling)
- `_types/` → `types/` (remove underscore prefix for public-facing module)

### 2. Move Core Module
- `mode.py` → `core/mode.py` (consolidate core functionality)

### 3. Backward Compatibility
- All changes include backward compatibility shims with deprecation warnings
- Existing imports continue to work with helpful migration messages
- Warnings guide users to new import paths for v2.0.0 migration

### 4. Import Updates
- Updated all internal imports to use new paths
- Provider client files now import Mode directly from core
- No breaking changes for external users

## Testing
- Existing tests pass with deprecation warnings (expected)
- Backward compatibility verified through test suite
- All changes preserve existing functionality

## Benefits
- Clearer module naming improves code discoverability
- Consolidation of core modules improves maintainability
- Easier for new contributors to understand codebase structure
- Prepares for clean v2.0.0 breaking change release
