+++ instructor/__init__.py
@@ -1,6 +1,6 @@
 import importlib.util
 
-from .mode import Mode
+from .core.mode import Mode
 from .processing.multimodal import Image, Audio
 
 from .dsl import (
