+++ instructor/_types/_alias.py
@@ -1,29 +1,24 @@
-from typing import Literal
+"""Backwards compatibility module for instructor._types._alias.
 
-from typing_extensions import TypeAlias
+This module has been renamed to types._alias for clarity.
+"""
 
-ModelNames: TypeAlias = Literal[
-    "gpt-4o",
-    "gpt-4-0125-preview",
-    "gpt-4-turbo-preview",
-    "gpt-4-1106-preview",
-    "gpt-4-vision-preview",
-    "gpt-4",
-    "gpt-4-0314",
-    "gpt-4-0613",
-    "gpt-4-32k",
-    "gpt-4-32k-0314",
-    "gpt-4-32k-0613",
-    "gpt-3.5-turbo",
-    "gpt-3.5-turbo-16k",
-    "gpt-3.5-turbo-0301",
-    "gpt-3.5-turbo-0613",
-    "gpt-3.5-turbo-1106",
-    "gpt-3.5-turbo-0125",
-    "gpt-3.5-turbo-16k-0613",
-    "gpt-3.5-turbo-instruct",
-    "text-embedding-ada-002",
-    "text-embedding-ada-002-v2",
-    "text-embedding-3-small",
-    "text-embedding-3-large",
-]
+import warnings
+
+
+def __getattr__(name: str):
+    """Lazy import to provide backward compatibility for _types._alias imports."""
+    warnings.warn(
+        f"Importing from 'instructor._types._alias' is deprecated and will be removed in v2.0.0. "
+        f"Please update your imports to use 'instructor.types._alias.{name}' instead",
+        DeprecationWarning,
+        stacklevel=2,
+    )
+
+    from ..types import _alias
+
+    # Try to get the attribute from the types._alias module
+    if hasattr(_alias, name):
+        return getattr(_alias, name)
+
+    raise AttributeError(f"module '{__name__}' has no attribute '{name}'")
