+++ instructor/dsl/__init__.py
@@ -2,7 +2,7 @@ from .iterable import IterableModel
 from .maybe import Maybe
 from .partial import Partial
 from .citation import CitationMixin
-from .simple_type import is_simple_type, ModelAdapter
+from .primitives import is_simple_type, ModelAdapter
 from . import validators  # Backwards compatibility module
 
 __all__ = [  # noqa: F405
