+++ instructor/providers/cohere/utils.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from ...mode import Mode
+from ...core.mode import Mode
 
 
 def reask_cohere_tools(
