+++ instructor/types/_alias.py
@@ -1,0 +1,29 @@
+from typing import Literal
+
+from typing_extensions import TypeAlias
+
+ModelNames: TypeAlias = Literal[
+    "gpt-4o",
+    "gpt-4-0125-preview",
+    "gpt-4-turbo-preview",
+    "gpt-4-1106-preview",
+    "gpt-4-vision-preview",
+    "gpt-4",
+    "gpt-4-0314",
+    "gpt-4-0613",
+    "gpt-4-32k",
+    "gpt-4-32k-0314",
+    "gpt-4-32k-0613",
+    "gpt-3.5-turbo",
+    "gpt-3.5-turbo-16k",
+    "gpt-3.5-turbo-0301",
+    "gpt-3.5-turbo-0613",
+    "gpt-3.5-turbo-1106",
+    "gpt-3.5-turbo-0125",
+    "gpt-3.5-turbo-16k-0613",
+    "gpt-3.5-turbo-instruct",
+    "text-embedding-ada-002",
+    "text-embedding-ada-002-v2",
+    "text-embedding-3-small",
+    "text-embedding-3-large",
+]
